Internal change
PiperOrigin-RevId: 331052608

+++ official/vision/beta/dataloaders/utils.py
@@ -23,11 +23,11 @@ from official.vision.beta.ops import preprocess_ops
 def process_source_id(source_id):
   """Processes source_id to the right format."""
   if source_id.dtype == tf.string:
-    source_id = tf.cast(tf.strings.to_number(source_id), tf.int32)
+    source_id = tf.cast(tf.strings.to_number(source_id), tf.int64)
   with tf.control_dependencies([source_id]):
     source_id = tf.cond(
         pred=tf.equal(tf.size(input=source_id), 0),
-        true_fn=lambda: tf.cast(tf.constant(-1), tf.int32),
+        true_fn=lambda: tf.cast(tf.constant(-1), tf.int64),
         false_fn=lambda: tf.identity(source_id))
   return source_id
